Re-introduce persistent of the list of repositories
This is now incorporated into LastCheck, so we always have a consistent state of the last known state, and we don't need to wait for anything to load (apart from the user name).

It also allows us to remove the dependency on the ReposGetResponse type in parts of the codebase that shouldn't need to be aware of the underlying GitHub type.