Make background script persistent
Notifications are often not getting delivered. I suspect this is because Chrome thinks that the extension is not doing anything (because it's waiting for a bunch of network requests to finish), gets impatient and kills it.

Making the background script persistent seems like the only option to fix this (happy to hear otherwise!).