Fix a bug where any comment on a PR would make it re-appear
We were relying on pr.updated_at, which actually gets updated any time there is any activity on the PR, including someone else commenting on it. It turns out pr.head.repo.pushed_at is also updated even when a different branch (master) is pushed, so we can't rely on that either.

This changes the approach to only notifying you when the author replies with at least one comment.

Further improvements could be made, for example looking at whether the author responded to you, or to another reviewer. For now, this should already decrease false positives significantly.