Remove optimisation to avoid loading comments and reviews unnecessarily (#233)
GitHub API's behaviour seems to have changed in the last week. Before, a PR's updated_at field would change whenever a comment or review was posted. This no longer seems to be the case, which means that our previous optimisation to skip reloading comments and reviews is now incorrect.

This also changes the refresh rate back to every 3 minutes, to reduce the chances of getting rate limited.

On the bright side, the code is now simpler :)