Add status and notifications for your own PRs
This introduces the distinction in your own PRs between:
- a PR that is pending review (with or without assigned reviewers)
- a PR where changes have been requested
- a PR that has been approved

A notification is shown when your PR is approved, or when new changes are requested. Note that no notification is shown when a new comment is posted, as this isn't necessarily a call to action (and we can't easily know whether it's a positive comment that can safely be ignored, or a negative comment that's equivalent to a request for change).

Additionally, the status of each PR is shown in the popup under the "My PRs" tab.

This fixes #160.