Sort PRs by most recent first (#304)
Until now, we were relying on GitHub's sorting for PRs, via the following query:
```typescript
        octokit.search.issuesAndPullRequests.endpoint.merge({
          q: `is:pr ${query}`,
          sort: "updated",
          order: "desc"
        })
```

However, this didn't always result in the expected order of PRs. For example, when someone posts a comment on a PR, it doesn't seem to bump it up in the list.

This replaces the sorting logic by manually sorting PRs instead, taking into account all comments, commits and reviews.