Add status and notifications for your own PRs (#303)
# Why

Until now, PR Monitor was primarily useful to review incoming PRs. However, it didn't give any feedback about your own PRs, e.g. when they were approved or "rejected".

# What

This introduces the distinction in your own PRs between:
- a PR that is pending review (with or without assigned reviewers)
- a PR where changes have been requested
- a PR that has been approved

A notification is shown when your PR is approved, or when new changes are requested. Note that no notification is shown when a new comment is posted, as this isn't necessarily a call to action (and we can't easily know whether it's a positive comment that can safely be ignored or a negative comment that's equivalent to a request for change).

Additionally, the status of each PR is shown in the popup under the "My PRs" tab.

# Implementation details

The following changes were made:
- Make an extra request to get PR details in order to fetch the list of requested reviewers, which unfortunately isn't included in the /issues/search endpoint.
- Add more granular `PullRequestStatus` values to distinguish between possible states.
- Add outgoingPullRequestStatus() which computes the status based on the PR information.
- Add badges to visually represent the status of each of your own PRs.
- Add notification handling of approved and "rejected" PRs.